+++ src/Config/Repository.php
@@ -78,6 +78,7 @@ final class Repository extends AbstractRepository
         return [
             ConfigFormGUI::KEY_ROLES                                      => [Config::TYPE_JSON, []],
             ConfigFormGUI::KEY_SHOW_ASSISTANTS                            => [Config::TYPE_BOOLEAN, false],
+            ConfigFormGUI::KEY_SHOW_DEPUTIES                              => [Config::TYPE_BOOLEAN, false],
             ConfigFormGUI::KEY_SHOW_ENROLMENT_WORKFLOW                    => [Config::TYPE_BOOLEAN, false],
             ConfigFormGUI::KEY_SHOW_EXCEL_IMPORT                          => [Config::TYPE_BOOLEAN, false],
             ConfigFormGUI::KEY_SHOW_EXCEL_IMPORT_CONFIG                   => [Config::TYPE_BOOLEAN, true],