+++ src/EnrolmentWorkflow/Repository.php
@@ -7,6 +7,7 @@ use srag\DIC\SrUserEnrolment\DICTrait;
 use srag\Plugins\SrUserEnrolment\Config\ConfigFormGUI;
 use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Action\Repository as ActionsRepository;
 use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Assistant\Repository as AssistantsRepository;
+use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Deputy\Repository as DeputiesRepository;
 use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Request\Repository as RequestsRepository;
 use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\Repository as RulesRepository;
 use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\SelectWorkflow\Repository as SelectedWorkflowsRepository;
@@ -74,12 +75,22 @@ final class Repository
 
 
     /**
+     * @return DeputiesRepository
+     */
+    public function deputies() : DeputiesRepository
+    {
+        return DeputiesRepository::getInstance();
+    }
+
+
+    /**
      * @internal
      */
     public function dropTables()/*: void*/
     {
         $this->actions()->dropTables();
         $this->assistants()->dropTables();
+        $this->deputies()->dropTables();
         $this->requests()->dropTables();
         $this->rules()->dropTables();
         $this->selectedWorkflows()->dropTables();
@@ -116,6 +127,7 @@ final class Repository
     {
         $this->actions()->installTables();
         $this->assistants()->installTables();
+        $this->deputies()->installTables();
         $this->requests()->installTables();
         $this->rules()->installTables();
         $this->selectedWorkflows()->installTables();