+++ src/EnrolmentWorkflow/Request/class.RequestInfoGUI.php
@@ -62,7 +62,7 @@ class RequestInfoGUI
     {
         $this->request = self::srUserEnrolment()->enrolmentWorkflow()->requests()->getRequestById(filter_input(INPUT_GET, self::GET_PARAM_REQUEST_ID));
 
-        if (!self::srUserEnrolment()->enrolmentWorkflow()->requests()->hasRequestAccess($this->request, self::dic()->user()->getId())) {
+        if ($this->single ? $this->request->getUserId() !== intval(self::dic()->user()->getId()) : false) {
             die();
         }